.scripts/requirements.txt
gql==3.4.1
PyGithub==1.59.1